lib/agents/enterprise-billing-system.ts
import { BaseAgent, AgentConfig } from './base';

export interface EnterpriseBillingSystemConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class EnterpriseBillingSystemAgent extends BaseAgent {
  private config: EnterpriseBillingSystemConfig;

  constructor(config: EnterpriseBillingSystemConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Enterprise Billing System implementation
    await new Promise(resolve => setTimeout(resolve, 1000));
    
    return `Enterprise Billing System completed successfully for: ${input}`;
  }

  async process(data: any): Promise<any> {
    // Advanced enterprise billing system processing
    return { status: 'success', data };
  }
}