__tests__/lib/agents/enterprise-billing-system.test.ts
import { EnterpriseBillingSystemAgent } from '@/lib/agents/enterprise-billing-system';

describe('Enterprise Billing SystemAgent', () => {
  let agent: EnterpriseBillingSystemAgent;

  beforeEach(() => {
    agent = new EnterpriseBillingSystemAgent({
      name: 'Test Enterprise Billing System Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      const result = await agent.execute('test input');
      expect(result).toContain('Enterprise Billing System completed successfully');
    });

    it('should handle empty input', async () => {
      const result = await agent.execute('');
      expect(result).toBeDefined();
    });
  });

  describe('process', () => {
    it('should process data correctly', async () => {
      const result = await agent.process({ test: 'data' });
      expect(result.status).toBe('success');
    });
  });
});